requirements.txt
pytorch>=1.10.0
transformers==4.8.1
nltk
wandb